+++ packages/cli/src/commands/diff/index.ts
@@ -5,6 +5,7 @@ export default defineCommand({
   subCommands: {
     apply: () => import('./apply.ts').then((m) => m.default),
     show: () => import('./show.ts').then((m) => m.default),
-    create: () => import('./create.ts').then((m) => m.default)
+    create: () => import('./create.ts').then((m) => m.default),
+    visual: () => import('./visual.ts').then((m) => m.default)
   }
 })